chore(ts): rename useMessageInfo to ts
