+++ src/components/MessagesList/MessagesGroup/Message/MessagePart/MessageBody.vue
@@ -147,7 +147,7 @@ import CallButton from '../../../../TopBar/CallButton.vue'
 import ConversationActionsShortcut from '../../../../UIShared/ConversationActionsShortcut.vue'
 import Poll from './Poll.vue'
 import { useIsInCall } from '../../../../../composables/useIsInCall.js'
-import { useMessageInfo } from '../../../../../composables/useMessageInfo.js'
+import { useMessageInfo } from '../../../../../composables/useMessageInfo.ts'
 import { CONVERSATION, MESSAGE } from '../../../../../constants.ts'
 import { hasTalkFeature } from '../../../../../services/CapabilitiesManager.ts'
 import { EventBus } from '../../../../../services/EventBus.ts'
