+++ src/components/MessagesList/MessagesGroup/MessagesGroup.vue
@@ -32,7 +32,7 @@ import { t } from '@nextcloud/l10n'
 import { computed, toRefs } from 'vue'
 import AvatarWrapper from '../../AvatarWrapper/AvatarWrapper.vue'
 import Message from './Message/Message.vue'
-import { useMessageInfo } from '../../../composables/useMessageInfo.js'
+import { useMessageInfo } from '../../../composables/useMessageInfo.ts'
 import { ATTENDEE, AVATAR } from '../../../constants.ts'
 import { useActorStore } from '../../../stores/actor.ts'
 import { useGuestNameStore } from '../../../stores/guestName.js'
