+++ src/components/Quote.vue
@@ -64,7 +64,7 @@ import PencilIcon from 'vue-material-design-icons/Pencil.vue'
 import AvatarWrapper from './AvatarWrapper/AvatarWrapper.vue'
 import DefaultParameter from './MessagesList/MessagesGroup/Message/MessagePart/DefaultParameter.vue'
 import FilePreview from './MessagesList/MessagesGroup/Message/MessagePart/FilePreview.vue'
-import { useMessageInfo } from '../composables/useMessageInfo.js'
+import { useMessageInfo } from '../composables/useMessageInfo.ts'
 import { ATTENDEE, AVATAR } from '../constants.ts'
 import { EventBus } from '../services/EventBus.ts'
 import { useActorStore } from '../stores/actor.ts'
