+++ src/composables/useMessageInfo.ts
